test: I think the most relevant thing for your question here would be to run terraform show to see Terraform’s current record of objects it’s tracking, and then compare that result with your configuration.
Terraform will only propose changes if the configuration is different than the remote objects. “No changes” as you reported earlier is the expected response if your current configuration matches the configuration of the remote objects.

If you can see a difference between the values shown by terraform show and the values your configuration represents (that is, the results of the expressions you wrote) then it would be helpful to share exactly which parts are different.